fix(events): fix subtle bug
In some very strange obscure edge cases, isolation with a call to DOMSource.events() can
cause a Destination's selector to be an empty string. This will throw an error that
Element.matchesSelector() can not be called with empty string. The fix, when the selector
is an empty string, test if the element is in isolation. If it is, continue to push the
event to its associated event subject.