feat(api-wrappers): implement functional api wrappers for DomSource
`elements(domSource: DomSource): Stream<Element>`

A functional implementation for `DomSource.elements()`.

`events(eventType: string, domSource: DomSource): Stream<Event>`

A functional implementation for `DomSource.events(eventType)`. This function is
curried by default.

`query(cssSelector: string, domSource: DomSource): DomSource`

A functional implementation for `DomSource.select(cssSelector)`. This function is
curried by default. The name of this function is `query` and not `select`
because it is a name conflict with the hyperscript helper function for the
`SELECT` HTML element.

`useCapture(domSource: DomSource): DomSource`

Combined with `events`, this allows for an equivalent of
`DomSource.events(eventType, { useCapture: true })`.